additive mace calculator
The code allows to calculate the energy and forces of two fragments within the same system
with two different calculators. The idea is to use mace calculator for pre-optimisation in surface 
reactions, where adsorbates can be represented by a seperate foundation model than the rest of the system. 
The same script may be used for supported nanocluster adsorption/reaction in a 3 calculator setup.